feat: Add Prisma module and service
This commit adds the Prisma module and service to the project. The Prisma module is imported in the app.module.ts file, along with the AuthModule and UserModule. The Prisma service is defined in the prisma.service.ts file and extends the PrismaClient class from the @prisma/client package. The Prisma service also implements the OnModuleInit interface and connects to the Prisma client in the onModuleInit method.